fix: change approach from importing dynamically to injecting dynamic import for dev i18n
