+++ packages/i18n/prepare-build.tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "dist/lib"
+    "outDir": "dist"
   },
   "include": ["lib/prepare_build.ts"]
 }
